Gmail/requirements.txt
composio-sdk
python-dotenv
supabase
pinecone-client

# wtf is this file doing why did chat create it...